+++ .env.example
@@ -2,6 +2,8 @@ PGHOST=localhost
 PGUSER=nano
 PGPASSWORD=shinonome
 PGDATABASE=kemonodb
+REDIS_HOST=kemono-redis
+REDIS_PORT=6379
 
 TELEGRAMTOKEN=
 TELEGRAMCHANNEL=
