+++ src/legacy.py
@@ -12,7 +12,7 @@ from shutil import move
 
 from PIL import Image
 from python_resumable import UploaderFlask
-from flask import Flask, jsonify, render_template, render_template_string, request, redirect, url_for, send_from_directory, make_response, g, abort, send_file, session, Blueprint
+from flask import Flask, jsonify, render_template, render_template_string, request, redirect, url_for, send_from_directory, make_response, g, abort, session, Blueprint
 from werkzeug.utils import secure_filename
 from slugify import slugify_filename
 import requests
@@ -22,58 +22,9 @@ from hashlib import sha256
 
 from .internals.database.database import get_cursor
 from .internals.utils.flask_cache import cache
-from .internals.utils.utils import make_cache_key
+from .internals.utils.utils import make_cache_key, relative_time, delta_key, allowed_file
 
 legacy = Blueprint('legacy', __name__)
-
-def delta_key(e):
-    return e['delta_date']
-
-def relative_time(date):
-    """Take a datetime and return its "age" as a string.
-    The age can be in second, minute, hour, day, month or year. Only the
-    biggest unit is considered, e.g. if it's 2 days and 3 hours, "2 days" will
-    be returned.
-    Make sure date is not in the future, or else it won't work.
-    Original Gist by 'zhangsen' @ https://gist.github.com/zhangsen/1199964
-    """
-
-    def formatn(n, s):
-        """Add "s" if it's plural"""
-
-        if n == 1:
-            return "1 %s" % s
-        elif n > 1:
-            return "%d %ss" % (n, s)
-
-    def qnr(a, b):
-        """Return quotient and remaining"""
-
-        return a / b, a % b
-
-    class FormatDelta:
-
-        def __init__(self, dt):
-            now = datetime.now()
-            delta = now - dt
-            self.day = delta.days
-            self.second = delta.seconds
-            self.year, self.day = qnr(self.day, 365)
-            self.month, self.day = qnr(self.day, 30)
-            self.hour, self.second = qnr(self.second, 3600)
-            self.minute, self.second = qnr(self.second, 60)
-
-        def format(self):
-            for period in ['year', 'month', 'day', 'hour', 'minute', 'second']:
-                n = getattr(self, period)
-                if n >= 1:
-                    return '{0} ago'.format(formatn(n, period))
-            return "just now"
-
-    return FormatDelta(date).format()
-
-def allowed_file(mime, accepted):
-    return any(x in mime for x in accepted)
 
 @legacy.route('/artists')
 @cache.cached(key_prefix=make_cache_key)
